refactor: 최종 게임 결과 출력 수정
