+++ src/main/java/bridge/message/OutputMessage.java
@@ -1,0 +1,5 @@
+package bridge.message;
+
+public class OutputMessage {
+    public static final String BRIDGE_FINAL = "최종 게임 결과";
+}
